SPRINT 1 SUBSTENTIAL 2
Responsible: Florian Beetz

Provided the foundation of the parser. The work done simplifies further implementation of the parser heavily.
Most of the interaction with the StAX API is implemented.
All classes are documented, however there are only few tests because this is basically an interface that can not be tested without implementation.